login validation

+++ database/migrations/2020_06_17_110149_create_categories_table.php
@@ -17,7 +17,7 @@ class CreateCategoriesTable extends Migration
             $table->id();
             $table->string('title');
             $table->string('description')->nullable();
-            $table->tinyInteger('parent_id')->unsigned()->default(0);
+            $table->integer('parent_id')->unsigned()->default(0);
             $table->timestamps();
         });
     }
